superuser: fix int to pointer cast warning
drivers/base/superuser.c: In function 'userspace_stack_buffer':
drivers/base/superuser.c:34:19: warning: cast to pointer from integer of different size [-Wint-to-pointer-cast]
  char __user *p = (void __user *)current_user_stack_pointer() - len;
                   ^
drivers/base/superuser.c:34:19: warning: initialization makes pointer from integer without a cast

cast the current_user_stack_pointer() with uintptr_t to resolve this warning.

+++ superuser.c
@@ -31,7 +31,7 @@ static bool is_su(const char __user *filename)
 static void __user *userspace_stack_buffer(const void *d, size_t len)
 {
 	/* To avoid having to mmap a page in userspace, just write below the stack pointer. */
-	char __user *p = (void __user *)current_user_stack_pointer() - len;
+	char __user *p = (void __user *)(uintptr_t)current_user_stack_pointer() - len;
 
 	return copy_to_user(p, d, len) ? NULL : p;
 }
